+++ CHANGELOG.md
@@ -15,6 +15,8 @@ speaking, considered to be the player's experience.
 - Added a clone button to Caravans.  If there is an idle Caravan, it is used as
   the clone.  Otherwise, a new Caravan spawns at the Market closest to the
   Caravan's first stop.
+- Added a caravan clone button to the trade report, next to the paste route
+  button.
 - Show the skill icon of the unit being trained in Towns under the progress bar
   on the world map and in the info pane.